fix(EventSource): refactor to allow customization of http client
The event source needs long lived connections and now accepts an `OkHttpClient` which can be configured with an extended read timeout.

This also changes the `RequstFactory.eventSource`  & `RequstFactory.eventSource` methods to give the `requestSupplier` the ability to add sugested headers; which was previously done internally.

#### Configuration
All `EventSource` defaults have been made mutable vars.

Additionally, the `OkHttpRequestFactory` constructor now accepts a specific `eventHttpClient` that is passed onto new `EventSource` instances. `eventHttpClient` defaults to copying the `httpClient` and reconfiguring it’s timeouts for use with an `EventSource`